Refactor: Post 관련 로직 분리, 경로 변경
