+++ src/components/UploadProfileForms/WellTalent.jsx
@@ -2,7 +2,7 @@
 import { css } from '@emotion/react';
 import { Typography } from '@material-ui/core';
 import palette from '../../lib/styles/palette';
-import TagBox from '../TagBox';
+import TagBox from '../common/TagBox';
 import { TagInput } from '../TextField';
 
 const WellTalent = ({ wellTalent, handleChangeFiled }) => {
@@ -28,12 +28,7 @@ const WellTalent = ({ wellTalent, handleChangeFiled }) => {
 
       <TagBox tagList={wellTalent} onClick={handleDeleteTalent} />
 
-      <TagInput
-        autoFocus
-        tagList={wellTalent}
-        onCreate={handleCreateTalent}
-        css={talentInput}
-      />
+      <TagInput autoFocus tagList={wellTalent} onCreate={handleCreateTalent} css={talentInput} />
     </>
   );
 };
