--- src/lib/util/postStatus.js
@@ -1,16 +1,0 @@
-const postStatus = {
-  Close: {
-    name: '',
-  },
-  Open: {
-    name: '채팅 오픈',
-  },
-  Reservation: {
-    name: '채팅 예약',
-  },
-  Notice: {
-    name: '공지',
-  },
-};
-
-export default postStatus;
