+++ src/lib/util/postType.js
@@ -1,0 +1,16 @@
+const postType = {
+    Close: {
+        name: '',
+    },
+    Open: {
+        name: '채팅 오픈',
+    },
+    Reservation: {
+        name: '채팅 예약',
+    },
+    Notice: {
+        name: '공지',
+    },
+};
+
+export default postType;
