+++ src/routers/ChannelRouter.jsx
@@ -6,26 +6,35 @@ import ChannelPostListPage from '../pages/channel/ChannelPostListPage';
 import ChannelPostPage from '../pages/channel/ChannelPostPage';
 import EditChannelPage from '../pages/channel/EditChannelPage';
 import CheckJoinChannel from '../middlewares/CheckJoinChannel';
-import ChannelPostWritingPage from '../pages/channel/ChannelPostWritingPage';
+import EditPostPage from '../pages/channel/EditPostPage';
 
 const ChannelRouter = () => {
   return (
     <>
       <Route path="/channel/:channelId/profile" exact component={ChannelProfilePage} />
       <Route
-        path="/channel/:channelId/home" exact
+        path="/channel/:channelId/home"
+        exact
         render={(props) => <CheckJoinChannel {...props} Component={ChannelHomePage} />}
       />
       <Route
-        path="/channel/:channelId/postList" exact
+        path="/channel/:channelId/postList"
+        exact
         render={(props) => <CheckJoinChannel {...props} Component={ChannelPostListPage} />}
       />
       <Route
-        path="/channel/:channelId/writing" exact
-        render={(props) => <CheckJoinChannel {...props} Component={ChannelPostWritingPage} />}
+        path="/channel/:channelId/chat"
+        exact
+        render={(props) => <CheckJoinChannel {...props} Component={ChannelPostListPage} />}
       />
       <Route
-        path="/channel/:channelId/post/:postId" exact
+        path="/channel/:channelId/writing"
+        exact
+        render={(props) => <CheckJoinChannel {...props} Component={EditPostPage} />}
+      />
+      <Route
+        path="/channel/:channelId/post/:postId"
+        exact
         render={(props) => <CheckJoinChannel {...props} Component={ChannelPostPage} />}
       />
       <Route path="/channel/edit" exact render={(props) => <EditChannelPage />} />
